add tests and creating deployment reconcile usecase
